Implementing Dial Generic
Dial generic is a class that encapsules common functionality present in all dial types. The plan is to have it cover both single rotation and repeat rotation knobs. In this commit, the Dial Generic class has been integrated with the Dial_full_motion dial type, with dial_regular to follow. For the time being until developed, most of the dial_snapto class is commented out to prevent build errors from WIP code.